+++ pkgs/intl4x/lib/src/first_day_of_week/first_day_of_week_stub.dart
@@ -2,10 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/// Support for Unicode locale identifiers.
-///
-/// Use the [Locale] class to represent and manipulate locales.
-/// @docImport 'locale.dart';
-library;
+import '../locale/locale.dart';
+import '../options.dart';
 
-export 'src/locale/locale.dart' show Locale;
+Weekday getFirstDayOfWeekECMA(Locale locale) =>
+    throw UnimplementedError('Cannot use ECMA outside of web environments.');